backend/requirements.txt
fastapi
uvicorn
pypdf
faiss-cpu
google-generativeai
python-multipart
pdfplumber
